+++ testsuite/test_arrayexpression.cpp
@@ -232,8 +232,8 @@ BOOST_FIXTURE_TEST_CASE( grid_constructor, ArrayExpressionTest )
   boost::timer::progress_display show_progress(4*N);
 
   for (int i=0; i<N; i++) {
-    Array<int,2> s1(idist_small(rGen), idist_small(rGen));
-    Array<int,2> s2(idist_small(rGen), idist_small(rGen));
+    Array<size_t,2> s1(idist_small(rGen), idist_small(rGen));
+    Array<size_t,2> s2(idist_small(rGen), idist_small(rGen));
     Grid<int,2> grid(s1+s2);
 
     BOOST_CHECK_EQUAL(grid.getLo()[0], 0);
